tickeiting_ms/requirements.txt
django==4.2
Pillow
djangorestframework